Move jobs into their own directory
