+++ lib/tooling_invoker/jobs/analyzer_job.rb
@@ -1,4 +1,4 @@
-module ToolingInvoker
+module ToolingInvoker::Jobs
   class AnalyzerJob < Job
     def initialize(id, *args)
       super(id, *args)
