+++ lib/tooling_invoker/jobs/representer_job.rb
@@ -1,4 +1,4 @@
-module ToolingInvoker
+module ToolingInvoker::Jobs
   class RepresenterJob < Job
     def initialize(id, *args)
       super(id, *args)
