+++ test/jobs/representer_job_test.rb
@@ -1,7 +1,7 @@
 require 'test_helper'
 
-module ToolingInvoker
-  class TestRunnerJobTest < Minitest::Test
+module ToolingInvoker::Jobs
+  class RepresenterJobTest < Minitest::Test
     def test_everything_is_set_correctly
       job_id = "213"
       language = "ruby"
@@ -10,7 +10,7 @@ module ToolingInvoker
       container_version = "v3"
       timeout = "10"
 
-      test_run = TestRunnerJob.new(job_id, language, exercise, s3_uri, container_version, timeout)
+      test_run = RepresenterJob.new(job_id, language, exercise, s3_uri, container_version, timeout)
       assert_equal job_id, test_run.id
       assert_equal language, test_run.language
       assert_equal exercise, test_run.exercise
@@ -23,8 +23,8 @@ module ToolingInvoker
         "/mnt/exercism-iteration/",
         "/mnt/exercism-iteration/"
       ], test_run.invocation_args
-      assert_equal ["results.json"], test_run.output_filepaths
-      assert_equal "/opt/test-runner", test_run.working_directory
+      assert_equal ["representation.txt", "mapping.json"], test_run.output_filepaths
+      assert_equal "/opt/representer", test_run.working_directory
     end
   end
 end
